Here's what I came up with for building the visibility graph. The function takes two arrays of adjacency-list pointers (one empty, one populated with the edges of the obstacles), arrays of x-values and y-values, and the number of total vertices (including the start and goal vertices, which are assumed to be the last two in the array).
The first test case (commented out) is:
      *         * G
    / |
   /  |    *----*
  /   |    |   /
 *----*    |  /
           | /
 * S       *

The function doesn't correctly handle overlapping objects. The uncommented
test case is:

           *        * G
         / |
      / *--+-------*
   /    |  |     /
 *------+--*  /
        |  /
 * S    *

more or less, and the output is the same. I think the bug is in the
intersect function, but I'm not sure. I will try and spend some time in
gdb...

Anyway, the functions this affects are
1. the one that reads the files: This needs to fill the obstacles array
with adjacency lists representing the obstacle edges, assign a value to
num_vertices, and fill the x and y arrays.
2. Dijkstra's algorithm; Not that there were many ways to do this, anyway.
If we search from the goal state, instead of the start, storing in an
int_list array d both the distance so far (d->v) and the node from which
that distance was obtained (d->next), d[START] will be a list of nodes
forming the shortest path with no additional processing.

Just some ideas, feedback is welcome.